# el-pasas/wwwroot/swagger-original.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "",
    "version" : "0.1",
    "title" : "eDocument API",
    "termsOfService" : ""
  },
  "host" : "el-pasas.azurewebsites.net",
  "basePath" : "/api",
...
  "security" : [ {
    "Token" : [ ]
  } ],
  "paths" : {
...
    "/drivingLicence/{userId}" : {
      "get" : {
        "tags" : [ "DrivingLicence" ],
        "summary" : "Get Driving Licence for the User",
        "operationId" : "getDrivingLicence",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successfully retrieved Driving Licence model",
            "schema" : {
              "$ref" : "#/definitions/DrivingLicence"
            }
          },
          "404" : {
            "description" : "Driving Licence for this {userId} not found"
          }
        }
      }
    },
    "/drivingLicence/{userId}/image/small" : {
      "get" : {
        "tags" : [ "DrivingLicence" ],
        "summary" : "Get small profile image for the Driving Licence which belongs to the User",
        "operationId" : "getDrivingLicenceImageSmall",
        "produces" : [ "image/jpeg" ],
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successfully retrieved Driving Licence small profile image"
          },
          "404" : {
            "description" : "Driving Licence small image for this {userId} not found"
          }
        }
      }
    },
    "/drivingLicence/{userId}/signature" : {
      "get" : {
        "tags" : [ "DrivingLicence" ],
        "summary" : "Get signature image for the Driving Licence which belongs to the User",
        "operationId" : "getDrivingLicenceSignature",
        "produces" : [ "image/jpeg" ],
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successfully retrieved Driving Licence signature image"
          },
          "404" : {
            "description" : "Driving Licence signature image for this {userId} not found"
          }
        }
      }
    },
    "/drivingLicence/{userId}/qr" : {
      "get" : {
        "tags" : [ "DrivingLicence" ],
        "summary" : "Get QR code for the Driving Licence which belongs to the User",
        "operationId" : "getDrivingLicenceQr",
        "produces" : [ "text/plain" ],
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successfully retrieved Driving Licence QR code"
          },
          "404" : {
            "description" : "Driving Licence QR code for this {userId} not found"
          }
        }
      }
    },
    "/drivingLicence/{userId}/nfc" : {
      "get" : {
        "tags" : [ "DrivingLicence" ],
        "summary" : "Get NFC code for the Driving Licence which belongs to the User",
        "operationId" : "getDrivingLicenceNfc",
        "produces" : [ "text/plain" ],
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successfully retrieved Driving Licence NFC code"
          },
          "404" : {
            "description" : "Driving Licence NFC code for this {userId} not found"
          }
        }
      }
    },
    "/drivingLicence/{userId}/image" : {
      "get" : {
        "tags" : [ "DrivingLicence" ],
        "summary" : "Get profile image for the Driving Licence which belongs to the User",
        "operationId" : "getDrivingLicenceImage",
        "produces" : [ "image/jpeg" ],
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successfully retrieved Driving Licence profile image"
          },
          "404" : {
            "description" : "Driving Licence image for this {userId} not found"
          }
        }
      },
      "put" : {
        "tags" : [ "DrivingLicence" ],
        "summary" : "Update profile image for the Driving Licence which belongs to the User",
        "operationId" : "putDrivingLicenceImage",
        "consumes" : [ "multipart/form-data" ],
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "image",
          "in" : "formData",
          "description" : "Image to upload",
          "required" : true,
          "type" : "file"
        } ],
        "responses" : {
          "204" : {
            "description" : "Successfully uploaded Driving Licence profile image"
          },
          "400" : {
            "description" : "Provided image is invalid"
          }
        }
      }
    },
    "/drivingLicence/{userId}/validate" : {
      "put" : {
        "tags" : [ "DrivingLicence" ],
        "summary" : "Request to revalidate and extend expiration for Driving Licence",
        "operationId" : "putDrivingLicenceValidation",
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "204" : {
            "description" : "Successfully requested Driving Licence revalidation"
          },
          "404" : {
            "description" : "Driving Licence for this {userId} not found"
          }
        }
      }
    },
...
  },
  "securityDefinitions" : {
    "Token" : {
      "type" : "basic"
    }
  },
  "definitions" : {
...
    "DrivingLicence" : {
      "type" : "object",
      "properties" : {
        "1_surname" : {
          "type" : "string",
          "example" : "Pavardenis"
        },
        "2_name" : {
          "type" : "string",
          "example" : "Vardenis"
        },
        "3_dateOfBirth" : {
          "type" : "string",
          "format" : "dd mm yyyy",
          "example" : "01 01 1985"
        },
        "4a_dateOfIssue" : {
          "type" : "string",
          "format" : "dd mm yyyy",
          "example" : "24 11 2016"
        },
        "4b_dateOfExpiry" : {
          "type" : "string",
          "format" : "dd mm yyyy",
          "example" : "24 11 2026"
        },
        "4c_issuedBy" : {
          "type" : "string",
          "example" : "VĮ Regitra"
        },
        "4d_personalNo" : {
          "type" : "integer",
          "example" : 31234567890
        },
        "5_licenceNo" : {
          "type" : "integer",
          "example" : 35983081
        },
        "categories" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Category"
          }
        }
      },
      "example" : {
        "4a_dateOfIssue" : "24 11 2016",
        "4c_issuedBy" : "VĮ Regitra",
        "2_name" : "Vardenis",
        "4b_dateOfExpiry" : "24 11 2026",
        "3_dateOfBirth" : "01 01 1985",
        "4d_personalNo" : 31234567890,
        "1_surname" : "Pavardenis",
        "5_licenceNo" : 35983081,
        "categories" : [ {
          "9_name" : "AM",
          "12_code" : "95.(24.11.21)",
          "11_dateOfExpiry" : "24.11.21",
          "10_dateOfIssue" : "19.01.13"
        }, {
          "9_name" : "AM",
          "12_code" : "95.(24.11.21)",
          "11_dateOfExpiry" : "24.11.21",
          "10_dateOfIssue" : "19.01.13"
        } ]
      }
    },
    "Category" : {
      "type" : "object",
      "properties" : {
        "9_name" : {
          "type" : "string",
          "enum" : [ "AM", "A1", "B1", "B", "D1", "D", "BE", "D1E", "DE" ]
        },
        "10_dateOfIssue" : {
          "type" : "string",
          "format" : "dd.mm.yy",
          "example" : "19.01.13"
        },
        "11_dateOfExpiry" : {
          "type" : "string",
          "format" : "dd.mm.yy",
          "example" : "24.11.21"
        },
        "12_code" : {
          "type" : "string",
          "example" : "95.(24.11.21)"
        }
      },
      "example" : {
        "9_name" : "AM",
        "12_code" : "95.(24.11.21)",
        "11_dateOfExpiry" : "24.11.21",
        "10_dateOfIssue" : "19.01.13"
      }
    },
...
  }
}